refactor(auth): enhance registration and profile creation flow
- Change numeric types from Double to Float in UserProfile and UserProfileEntity for consistency
- Add detailed logging and stepwise process in AuthRepository.register and registerWithGoogle methods
- Ensure existing sessions are deleted before new registration begins
- Implement immediate auto-login after registration for smoother user experience
- Store referral code securely if provided during registration
- Create user profile after successful registration and log detailed profile info
- Handle profile creation failures gracefully; defer resolution to ProfileViewModel
- Add createUserProfileIfNotExists method to AuthRepository to ensure profile existence
- Adjust UserRepository queries to expect list of userIds and referral codes for compatibility
- Update UserRepository.createUserProfile to include current timestamp fields and correct data types
- Modify UserRepository to handle userId and referralCode fields that may be lists or strings
- Remove deprecated coinsPerSecond property from UserProfile and related mapping functions
- Fix numeric literals in various places to Float type for consistency
- Improve error and success logs in ProfileViewModel when loading and creating profiles
- Optimize logout method with better logging and ensure local data is cleared regardless of errors
- Simplify logout and sign out handlers in UI navigation by removing unnecessary coroutine scopes
- Fix MiningScreenStats to display autoMiningRate instead of deprecated coinsPerSecond
- Refine updateUserProfileBalance method in MiningViewModel to accept Float parameter
- Update Preview data objects with Float literals for numeric fields
- Add consistent and clearer logging markers throughout authentication and profile-related code paths

+++ ktMobileApp/app/src/main/java/com/ekehi/network/presentation/viewmodel/MiningViewModel.kt
@@ -255,7 +255,7 @@ class MiningViewModel @Inject constructor(
                 // Update user profile immediately
                 val currentProfile = _userProfile.value
                 if (currentProfile is Resource.Success) {
-                    val newBalance = currentProfile.data.totalCoins + _sessionReward.value
+                    val newBalance = currentProfile.data.totalCoins + _sessionReward.value.toFloat()
                     updateUserProfileBalance(newBalance)
                 }
                 
@@ -396,9 +396,10 @@ class MiningViewModel @Inject constructor(
     }
     
     /**
-     * Updates the user profile with new balance
+     * Updates the user profile balance in the UI (does not persist to database)
+     * @param newBalance The new balance to display
      */
-    fun updateUserProfileBalance(newBalance: Double) {
+    fun updateUserProfileBalance(newBalance: Float) {
         val currentProfile = _userProfile.value
         if (currentProfile is Resource.Success) {
             val profile = currentProfile.data
